+++ src/1-one/change/api/index.js
@@ -5,8 +5,9 @@ import remove from './remove.js'
 import whitespace from './whitespace.js'
 import sort from './sort.js'
 import concat from './concat.js'
+import harden from './harden.js'
 
-const methods = Object.assign({}, caseFns, insert, replace, remove, whitespace, sort, concat)
+const methods = Object.assign({}, caseFns, insert, replace, remove, whitespace, sort, concat, harden)
 
 const addAPI = function (View) {
   Object.assign(View.prototype, methods)